Add files via upload
Welcome to the repository for Calculator Web Apps! This project is an HTML, CSS, and JavaScript-based calculator application that is both fully functional and user-friendly. Its elegant appearance and simple interface make it easier than ever to execute simple to complex mathematical computations. This web tool has you covered whether you're a student in need of rapid calculations or someone seeking for a handy internet calculator.

Key characteristics:
-User-Friendly Interface: The calculator has an intuitive user interface that is similar to a physical calculator, making it simple to use and easily recognisable for everyone.
-Basic and Advanced Operations: With only a few clicks, perform common mathematical operations like addition, subtraction, multiplication, and division.